hotfixt: 새벽 알파테스트 회의 반영

+++ src/main/java/com/example/modam/domain/bookcase/Application/BookCaseService.java
@@ -67,6 +67,9 @@ public class BookCaseService {
             throw new ApiException(ErrorDefine.BOOKCASE_NOT_FOUND);
         }
         BookCaseEntity data = optionalData.get();
+        if (data.getStatus() == BookState.AFTER) {
+            throw new ApiException(ErrorDefine.UNAUTHORIZED_STATUS);
+        }
 
         bookCaseRepository.delete(data);
     }
